Reorganize c_src directory tree
This commit only moves source files around to adopt compartmentalization of
related code together by directory.

The Makefile was update to reflect the new directory tree.
