+++ c_src/device/bcm.c
@@ -22,11 +22,11 @@
 #include <EGL/eglext.h>
 
 #define NANOVG_GLES2_IMPLEMENTATION
-#include "../nanovg/nanovg.h"
-#include "../nanovg/nanovg_gl.h"
+#include "nanovg.h"
+#include "nanovg_gl.h"
 
-#include "../types.h"
-#include "../comms.h"
+#include "scenic_types.h"
+#include "comms.h"
 #include "device.h"
 
 #define DEFAULT_SCREEN    0
@@ -295,7 +295,7 @@ char* device_gl_error() {
     case GL_STACK_OVERFLOW:
       return "GL_STACK_OVERFLOW";
 #endif
-      
+
     case GL_INVALID_FRAMEBUFFER_OPERATION:
       return "GL_INVALID_FRAMEBUFFER_OPERATION";
