+++ c_src/image/stb_image.h
