+++ c_src/scenic/script.c
