+++ c_src/scenic/utils.h
